% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calibration_year_index}
\alias{calibration_year_index}
\title{Synthetic Years Index}
\format{
An object of class \code{character} of length 20.
}
\usage{
calibration_year_index
}
\description{
We chose proxy years for the calibration time period of 1998-2017
by selecting the year from the 1980-1999 model inputs that most closely matched the
\href{https://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST}{DWR water year index}
in both the San Joaquin and Sacramento Basins of the targeted calibration year.
}
\keyword{datasets}
