% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mean_escapement_2013_2017}
\alias{mean_escapement_2013_2017}
\title{Mean Escapement 2013-2017}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31 rows and 5 columns.
}
\usage{
mean_escapement_2013_2017
}
\description{
Mean Grandtab in-river escapement values for CVPIA SIT watersheds (2013-2017)
}
\details{
see \link{grandtab} for more information on Grandtab
}
\keyword{datasets}
