% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_synth_years.R
\name{set_synth_years}
\alias{set_synth_years}
\title{Set Synthetic Year Series}
\usage{
set_synth_years(species = c("fr", "lfr", "wr", "sr", "st"))
}
\arguments{
\item{species}{either "fr", "lfr", "wr", "sr" for fall run, late fall run, winter run, or spring run respectively}
}
\description{
Generates model params for running the life cycle models in calibration mode
}
\details{
This function modifies the baseline model params for a run using
the synthetic year series. More details about the synthetic year series can be found
here: \code{\link{calibration_year_index}}.

Results of running \code{set_synth_years} for each run are cached and accessible via
\code{\link{fall_run_calibration}}, \code{\link{winter_run_calibration}},and
\code{\link{spring_run_calibration}}
}
\examples{
\dontrun{
params <- set_synth_years("fr")
fallRunDSM::fall_run_model(mode = "calibrate", seeds = seeds, ..params = params)
}
}
