% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{year_indicies}
\alias{year_indicies}
\alias{calibration_year_index}
\alias{calibration_year_spawn_index}
\alias{calibration_year_index_2019}
\alias{calibration_year_spawn_index_2019}
\title{Synthetic Years Index}
\usage{
calibration_year_index

calibration_year_spawn_index

calibration_year_index_2019

calibration_year_spawn_index_2019
}
\description{
We chose proxy years for the calibration time period of 1998-2017
by selecting the year from the 1980-1999 model inputs that most closely matched the
\href{https://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST}{DWR water year index}
in both the San Joaquin and Sacramento Basins of the targeted calibration year.
}
\details{
The 2019 calibration used a different selection of years using a similar methodology
}
\keyword{datasets}
