% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calibration_data}
\alias{calibration_data}
\alias{fall_run_calibration_params}
\title{Calibration Data}
\usage{
fall_run_calibration_params
}
\description{
Cached data used for calibration of DSM life cycle models
generated with \code{\link{set_synth_years}}
}
\keyword{datasets}
