% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calibration_data}
\alias{calibration_data}
\alias{fall_run_calibration}
\alias{spring_run_calibration}
\alias{winter_run_calibration}
\title{Calibration Data}
\format{
An object of class \code{list} of length 25.

An object of class \code{list} of length 26.

An object of class \code{list} of length 25.
}
\usage{
fall_run_calibration

spring_run_calibration

winter_run_calibration
}
\description{
data used for calibration of DSM life cycle models.
}
\keyword{datasets}
