% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_synth_years.R
\name{set_synth_years}
\alias{set_synth_years}
\title{Set Synthetic Year Series}
\usage{
set_synth_years(
  params,
  spawn_years = DSMCalibrationData::calibration_year_spawn_index,
  years = DSMCalibrationData::calibration_year_index
)
}
\arguments{
\item{params}{list object containing all arguments and coefficents for a life cycle model}

\item{spawn_years}{named array of synthetic year series for spawning inputs}

\item{years}{named array of synthetic year series for all other inputs}
}
\description{
Generates model params for running the life cycle models in calibration mode
}
\details{
This function modifies the baseline model params for a run using
the synthetic year series. More details about the synthetic year series can be found
here: \code{\link{calibration_year_index}}.

Results of running \code{set_synth_years} for each run are cached and accessible via
\code{\link{fall_run_calibration}}, \code{\link{winter_run_calibration}},and
\code{\link{spring_run_calibration}}
}
\examples{
\dontrun{
params <- set_synth_years(fallRunDSM::params)
fallRunDSM::fall_run_model(mode = "calibrate", seeds = seeds, ..params = params)
}
}
